pyproject.toml
[project]
name = "arxiv-scraper"
version = "0.1.0"
description = "A AI-powered arXiv scraper that fetches and summarizes research papers."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.11",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0"
]
